requirements.txt
selenium==3.141.0
Flask==2.0.1
Werkzeug==2.0.1
slackeventsapi==3.0.0
slackbot==1.0.0
slackclient==2.9.3
gunicorn==20.1.0
pytz
python-dotenv